perf(proto): minimize allocation of ProtocolReader
ProtocolReader uses an internal buffer for reading, but it is constantly
shrinked from the start side of the slice by `advance()`,
so the underlying array isnot reused.
When the intenal buffer run out of space, the `fill()` will make a new
slice. This adds GC pressure when reading audio streams.

This commit does the following to reduce memory allocation.

- Wrap the underlying *net.Conn with a bufio.Reader. It has a
  fixed-size buffer so the number of syscalls could be reduced.
  It also reuses its buffer so there is no need to handle that
  ourselves. bufio.Reader also comes with handy methods to deal with
  null-terminated strings.

- The reader keep two pieces of memory. One is a [8]byte for temporary storing
  integral types. The other is a bytes.Buffer for variable length items
  like strings and byte arrays. This is always reset when reading new
  data, so new memory is only allocated when reading in things larger than its
  size.