Fix data races in stream.state
PlaybackStream and RecordStream had both a 'state' variable that got
read and updated from different goroutines. This is a data race, and
shows up when running with the -race flag.

I've tried to make the code a bit less racy, and used an atomic value to
fix the warnings about the race condition. I'm still not 100% convinced
this fixes all races though.